+++ modules/base/default.nix
@@ -1,11 +1,9 @@
-{ inputs, ... }:
+{ ... }:
 rec
 {
-  imports = with inputs.self.nixosModules; [ apps gui hardware nix scripts secrets shell user ]
-  ++
+  imports =
   [
     ./console.nix
     ./firmware.nix
-    ./path.nix
   ];
 }