+++ path.nix
@@ -1,0 +1,4 @@
+{
+  system = "/etc/nixos";
+  keys = "/etc/nixos/secrets/unencrypted/gpg";
+}